removed basic etch script. fixed smart contract. synergetic contract cannot work until it is possible to package up model, data, and label into one object for passing to createProblem
