--- ml_examples/01_etch_sequential_model/model.etch
@@ -1,44 +1,0 @@
-//------------------------------------------------------------------------------
-//
-//   Copyright 2019 Fetch.AI Limited
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//       http://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License.
-//
-//------------------------------------------------------------------------------
-
-@action
-function main()
-
-    var data = readCSV(System.Argv(1));
-    var label = readCSV(System.Argv(2));
-    var test_data = readCSV(System.Argv(3));
-    var test_label = readCSV(System.Argv(4));
-
-    // set up a model architecture
-    var model = Model("sequential");
-    model.add("dense", 13u64, 10u64, "relu");
-    model.add("dense", 10u64, 10u64, "relu");
-    model.add("dense", 10u64, 1u64);
-    model.compile("mse", "adam");
-
-    // train the model
-    var batch_size = 10u64;
-    model.fit(data, label, batch_size);
-
-    // evaluate performance
-    var loss = model.evaluate();
-
-    // make predictions on all test data
-    var predictions = model.predict(test_data);
-
-endfunction
