+++ ml_examples/01_smart_contract_model/model.etch
